Change Skip List implementation.
oset_create|destroy: Add implementation.

node_create|destroy: Add implementation.

+++ modules/SkipList/ADTOrderedSet.c
@@ -1,15 +1,27 @@
 #include "ADTOrderedSet.h"
 
-#include <assert.h>
-#include <stdbool.h>
-#include <stdlib.h>
+#include <assert.h>   // assert
+#include <stdbool.h>  // true, false
+#include <stdlib.h>   // malloc, free, sizeof
 #include <string.h>
+
+/// @brief Levels of forward pointers a node can have.
+///
+/// Start at 16 levels, because the Oredered Set can contain 2^16 elements which is plenty.
+/// In the event the Oredered Set has 2^16 elements, double its max level.
+///
+#define OSET_LEVELS 16
 
 struct ordered_set {
     CompareFunc compare;
     DestroyFunc destroy_key;
     DestroyFunc destroy_value;
-    int size;
+
+    int max_level;  // Maximum level possible for a node of the Ordered Set.
+                    // Oredered Set can have 2^max_level, if that number is reached, then
+                    // double the max_level.
+
+    int size;  // Number of elements in the Ordered Set.
 
     OrderedSetNode first;
     OrderedSetNode last;
@@ -18,45 +30,43 @@ struct ordered_set {
 };
 
 struct ordered_set_node {
-    OrderedSetNode top;
-    OrderedSetNode bottom;
-    OrderedSetNode next;
+    OrderedSetNode* forward;
     OrderedSetNode previous;
 
-    bool is_header;
-    int level;
+    int* width;      // Number of bottom layer links being traversed by each of the forward links.
+    int levels;      // Number of forward links.
+    bool is_header;  // true, if node is header, otherwise false.
 
     void* key;
     void* value;
 };
 
-enum coin {
-    TAILS,
-    HEADS
-};
-
-static enum coin coinflip() {
-    return rand() % 2 ? HEADS : TAILS;
-}
-
 /// @brief Creates and returns an Ordered Set node.
 ///
-/// @param level The height of the node.
+/// @param levels Number of forward links.
 /// @param is_header true if node is header node, else false.
 ///
-static OrderedSetNode node_create(void* key, void* value, int level, bool is_header) {
+static OrderedSetNode node_create(void* key, void* value, int levels, bool is_header) {
     OrderedSetNode node = malloc(sizeof(*node));
     if (node == NULL) {
         return NULL;
     }
 
-    node->top = OSET_EOF;
-    node->bottom = OSET_EOF;
-    node->next = OSET_EOF;
-    node->previous = OSET_EOF;
+    // Allocate forward array and initialize it to NULL.
+    node->forward = calloc(levels * sizeof(*node->forward), levels);
+    if (node->forward == NULL) {
+        return NULL;
+    }
 
+    // Allocate width array.
+    node->width = malloc(levels * sizeof(*node->width));
+    if (node->width == NULL) {
+        return NULL;
+    }
+
+    node->levels = levels;
+    node->previous = OSET_BOF;
     node->is_header = is_header;
-    node->level = level;
 
     node->key = key;
     node->value = value;
@@ -69,101 +79,11 @@ static OrderedSetNode node_create(void* key, void* value, int level, bool is_hea
 /// Any operation on the Ordered Set node after its destruction, results in undefined behaviour.
 ///
 static void node_destroy(OrderedSetNode node, DestroyFunc destroy_key, DestroyFunc destroy_value) {
-    if (!node->is_header) {
-        if (node->level == 0) {
-            if (destroy_key != NULL) {
-                destroy_key(node->key);
-            }
-            if (destroy_value != NULL) {
-                destroy_value(node->value);
-            }
-        }
-    }
-
+    if (destroy_key != NULL) destroy_key(node->key);
+    if (destroy_value != NULL) destroy_value(node->value);
+    free(node->forward);
+    free(node->width);
     free(node);
-}
-
-static OrderedSetNode node_find_previous(OrderedSet oset, void* key) {
-    assert(key != NULL);
-
-    OrderedSetNode target = OSET_EOF;  // Points to previous node of node with node->key < key.
-
-    // Traverse from top to botom level.
-    OrderedSetNode node = oset->header;
-    while (node != OSET_EOF) {
-        // Traverse nodes in level with node->key < key.
-        while (node->next != OSET_EOF && oset->compare(node->next->key, key) < 0) {
-            node = node->next;
-        }
-
-        target = node;
-
-        node = node->bottom;
-    }
-
-    return target;
-}
-
-static void level_create(OrderedSet oset) {
-    OrderedSetNode new_header = node_create(
-        OSET_BOF, OSET_BOF, 1 + oset->header->level, true);
-
-    OrderedSetNode old_header = oset->header;
-
-    // Connect new_header with old_header verticaly.
-    new_header->bottom = old_header;
-    old_header->top = new_header;
-
-    // Update Ordered Set header.
-    oset->header = new_header;
-}
-
-static void node_promote(OrderedSet oset, OrderedSetNode node) {
-    OrderedSetNode target = node->previous;
-    while (target != OSET_BOF && target->top == OSET_EOF) {
-        target = target->previous;
-    }
-    if (target == OSET_BOF) {
-        level_create(oset);
-        target = oset->header;
-    } else {
-        target = target->top;
-    }
-
-    OrderedSetNode new_node = node_create(
-        node->key, node->value, 1 + node->level, node->is_header);
-
-    // Insert new_node after target.
-    new_node->next = target->next;
-    new_node->previous = target;
-    target->next = new_node;
-    if (new_node->next != OSET_EOF) {
-        new_node->next->previous = new_node;
-    }
-
-    // Connect new_node and node verticaly.
-    node->top = new_node;
-    new_node->bottom = node;
-}
-
-static void node_destroy_top(OrderedSetNode node, DestroyFunc destroy_key, DestroyFunc destroy_value) {
-    // Traverse nodes from bottom to top.
-    OrderedSetNode top = OSET_EOF;
-    while (node != OSET_EOF) {
-        top = node->top;
-
-        // Connect next and previous nodes of specified node directly.
-        if (node->next != OSET_EOF) {
-            node->next->previous = node->previous;
-        }
-        if (node->previous != OSET_BOF) {
-            node->previous->next = node->next;
-        }
-
-        node_destroy(node, destroy_key, destroy_value);
-
-        node = top;
-    }
 }
 
 OrderedSet oset_create(CompareFunc compare, DestroyFunc destroy_key, DestroyFunc destroy_value) {
@@ -176,12 +96,14 @@ OrderedSet oset_create(CompareFunc compare, DestroyFunc destroy_key, DestroyFunc
     oset->destroy_key = destroy_key;
     oset->destroy_value = destroy_value;
 
-    oset->first = OSET_EOF;
-    oset->last = OSET_EOF;
+    oset->max_level = OSET_LEVELS;
     oset->size = 0;
 
+    oset->first = OSET_EOF;
+    oset->last = OSET_EOF;
+
     // Header nodes don't need to have neither keys nor values.
-    oset->header = node_create(OSET_BOF, OSET_BOF, 0, true);
+    oset->header = node_create(OSET_BOF, OSET_BOF, oset->max_level, true);
     if (oset->header == NULL) {
         return OSET_ERROR;
     }
@@ -191,19 +113,15 @@ OrderedSet oset_create(CompareFunc compare, DestroyFunc destroy_key, DestroyFunc
 
 void oset_destroy(OrderedSet oset) {
     OrderedSetNode node = oset->header;
+
     while (node != OSET_EOF) {
-        OrderedSetNode bottom = node->bottom;
+        OrderedSetNode next = node->forward[0];
 
-        while (node != OSET_EOF) {
-            OrderedSetNode next = node->next;
+        node_destroy(node, oset->destroy_key, oset->destroy_value);
 
-            node_destroy(node, oset->destroy_key, oset->destroy_value);
-
-            node = next;
-        }
-
-        node = bottom;
+        node = next;
     }
+
     free(oset);
 }
 
@@ -223,65 +141,12 @@ int oset_size(OrderedSet oset) { return oset->size; }
 
 void oset_insert(OrderedSet oset, void* key, void* value) {
     assert(key != NULL);
-
-    OrderedSetNode node = node_find_previous(oset, key);
-
-    OrderedSetNode new_node = node_create(key, value, 0, false);
-
-    // Insert new_node after node.
-    new_node->next = node->next;
-    new_node->previous = node;
-    node->next = new_node;
-    if (new_node->next != OSET_EOF) {
-        new_node->next->previous = new_node;
-    }
-
-    // Promote new_node into top levels.
-    OrderedSetNode promotion = new_node;
-    while (coinflip() == HEADS) {
-        node_promote(oset, promotion);
-        promotion = promotion->top;
-    }
-
-    // Update first pointer.
-    if (oset->first == OSET_EOF || oset->compare(new_node->key, oset->first->key) < 0) {
-        oset->first = new_node;
-    }
-
-    // Update last pointer.
-    if (oset->last == OSET_EOF || oset->compare(new_node->key, oset->last->key) > 0) {
-        oset->last = new_node;
-    }
-
-    // Update size.
-    oset->size++;
 }
 
 bool oset_remove(OrderedSet oset, void* key) {
     assert(key != NULL);
 
-    OrderedSetNode node = node_find_previous(oset, key);
-    if (node->next == OSET_EOF || oset->compare(node->next->key, key) != 0) {
-        return false;
-    }
-
-    // Update first pointer.
-    if (node->next == oset->first) {
-        oset->first = oset->first->next;
-    }
-
-    // Update last pointer.
-    if (node->next == oset->last) {
-        oset->last = node->is_header ? OSET_BOF : node;
-    }
-
-    // Destroy node including its top levels.
-    node_destroy_top(node->next, oset->destroy_key, oset->destroy_value);
-
-    // Update size.
-    oset->size--;
-
-    return true;
+    return false;
 }
 
 void* oset_find(OrderedSet oset, void* key) {
@@ -303,11 +168,11 @@ void oset_concat(OrderedSet a, OrderedSet b) {}
 OrderedSetNode oset_find_node(OrderedSet oset, void* key) {
     assert(key != NULL);
 
-    OrderedSetNode node = node_find_previous(oset, key);
+    // OrderedSetNode node = node_find_previous(oset, key);
 
-    if (node->next != OSET_EOF && oset->compare(node->next->key, key) == 0) {
-        return node->next;
-    }
+    // if (node->next != OSET_EOF && oset->compare(node->next->key, key) == 0) {
+    //     return node->next;
+    // }
 
     return OSET_EOF;
 }
@@ -330,7 +195,8 @@ OrderedSetNode oset_last(OrderedSet oset) { return oset->last; }
 
 OrderedSetNode oset_next(OrderedSet oset, OrderedSetNode node) {
     assert(node != NULL);
-    return node->next;
+    // return node->next;
+    return NULL;
 }
 
 OrderedSetNode oset_previous(OrderedSet oset, OrderedSetNode node) {
